+++ server/component/style/conditionFailed/ConditionFailedView.php
@@ -1,0 +1,50 @@
+<?php
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+?>
+<?php
+require_once __DIR__ . "/../StyleView.php";
+
+/**
+ * The view class of the ConditionFailed style component.
+ * A ConditionFailed style is a container that allows to wrap content into a div tag that is shown only if the condition fails in conditionalContainer.
+ */
+class ConditionFailedView extends StyleView
+{
+    /* Constructors ***********************************************************/
+
+    /**
+     * The constructor.
+     *
+     * @param object $model
+     *  The model instance of the component.
+     */
+    public function __construct($model)
+    {
+        parent::__construct($model);
+    }
+
+    /* Public Methods *********************************************************/
+
+    /**
+     * Render the style view.
+     */
+    public function output_content()
+    {
+        require __DIR__ . "/tpl_conditionFailed.php";
+    }
+
+    /**
+     * Render the style view.
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_content_entry($entry_value)
+    {
+        $entry_data = $entry_value;
+        require __DIR__ . "/tpl_conditionFailed_entryValue.php";
+    }
+	
+}
+?>