+++ server/component/style/conditionalContainer/ConditionalContainerView.php
@@ -108,8 +108,12 @@ class ConditionalContainerView extends StyleView
                 var_dump($res);
             echo "</pre>";
         }
-        if($this->model->is_cms_page() || $res['result'])
+        if(
+        $this->model->is_cms_page() || $res['result']) {
             require __DIR__ . "/tpl_container.php";
+        } else {
+            require __DIR__ . "/tpl_failed_container.php";
+        }
     }
 
     /**
@@ -133,8 +137,11 @@ class ConditionalContainerView extends StyleView
                 var_dump($res);
             echo "</pre>";
         }
-        if($this->model->is_cms_page() || $res['result'])
+        if ($this->model->is_cms_page() || $res['result']) {
             require __DIR__ . "/tpl_container_entryValue.php";
+        } else {
+            require __DIR__ . "/tpl_failed_container_entryValue.php";
+        }
     }
 
     public function output_content_mobile()
@@ -164,6 +171,93 @@ class ConditionalContainerView extends StyleView
         $res = $this->model->compute_condition($this->condition);
         if ($res['result']) {
             return parent::output_content_mobile_entry($entry_value);
+        }
+    }
+
+    /**
+     * Render the content of all children of this view instance.
+     * Overwrite the basic function as we do not show the style which are in conditionFailed
+     */
+    public function output_children()
+    {
+        foreach ($this->children as $child) {
+            if ( $child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+                if($child->get_view() instanceof ConditionFailedView){
+                    // do nothhing. This child is shown only if the condition fails
+                }else{
+                    $child->output_content();
+                }
+            } else {
+                echo "invalid child element of type '" . gettype($child) . "'";
+            }
+        }
+    }
+
+    /**
+     * Render the content of all children of this view instance as entries
+     * * Overwrite the basic function as we do not show the style which are in conditionFailed
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_children_entry($entry_data)
+    {
+        foreach ($this->children as $child) {
+            if ($child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+                if ($child->get_view() instanceof ConditionFailedView) {
+                    // do nothhing. This child is shown only if the condition fails
+                } else {
+                    if (method_exists($child, 'output_content_entry')) {
+                        $child->output_content_entry($entry_data);
+                    } else {
+                        $child->output_content();
+                    }
+                }
+            } else {
+                echo "invalid child element of type '" . gettype($child) . "'";
+            };
+        }
+    }
+
+    /**
+     * Render the content conditionFailed
+     */
+    public function output_failed_children()
+    {
+        foreach ($this->children as $child) {
+            if ( $child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+                if($child->get_view() instanceof ConditionFailedView){
+                    $child->output_content();
+                }else{
+                    // do nothhing condition failed
+                }
+            } else {
+                echo "invalid child element of type '" . gettype($child) . "'";
+            }
+        }
+    }
+
+    /**
+     * Render the content of conditionfailed
+     * @param array $entry_value
+     * the data for the entry value
+     */
+    public function output_failed_children_entry($entry_data)
+    {
+        foreach ($this->children as $child) {
+            if ($child instanceof StyleComponent || $child instanceof BaseStyleComponent) {
+                if ($child->get_view() instanceof ConditionFailedView) {
+                    
+                    if (method_exists($child, 'output_content_entry')) {
+                        $child->output_content_entry($entry_data);
+                    } else {
+                        $child->output_content();
+                    }
+                } else {
+                    // do nothhing condition failed
+                }
+            } else {
+                echo "invalid child element of type '" . gettype($child) . "'";
+            };
         }
     }
 }